requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
psycopg2-binary>=2.9
python-decouple>=3.8
bcrypt>=4.0
drf-spectacular>=0.26.2
django-cors-headers>=4.3.1
Pillow>=11.2.1
pandas>=2.2.1
scikit-learn>=1.6.1
numpy>=1.26.4
joblib>=1.3.2
